Release v4.6.8 [master]

See merge request desktop/desktop!1378

+++ src/MEGASync/google_breakpad/client/linux/minidump_writer/cpu_set.h
@@ -54,7 +54,7 @@ public:
   // Parse a sysfs file to extract the corresponding CPU set.
   bool ParseSysFile(int fd) {
     char buffer[512];
-    int ret = sys_read(fd, buffer, sizeof(buffer)-1);
+    int ret = static_cast<int>(sys_read(fd, buffer, sizeof(buffer)-1));
     if (ret < 0)
       return false;
